Move automodeling logic to AutoModeler class
